Add admin functionality for user management and document handling
