+++ admin/VerDocumentos.php
@@ -1,13 +1,11 @@
 <?php
 session_start();
-
 if (!isset($_SESSION['username'])) {
-    header("Location: ../login/login.html");
+    header('Location: /PORTALTRANSPARENCIAMUNI/login/login.html');
     exit();
 }
-?>
-<?php
-require_once '../connection/db.php';
+
+require_once __DIR__ . '/../connection/db.php';
 
 // Obtener los tipos activos desde la base de datos
 $sql = "SELECT prefijo, nombre FROM tipos WHERE estado = 'activo'";
@@ -71,8 +69,8 @@ $documentos = $stmt->fetchAll(PDO::FETCH_ASSOC);
     <title>Documentos Subidos</title>
     <link rel="icon" href="../public/img/logo_white.ico" type="image/x-icon">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/ver_documentos.css">
+    <link rel="stylesheet" href="../public/css/style.css">
+    <link rel="stylesheet" href="../public/css/ver_documentos.css">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
