+++ admin/editar_documento.php
@@ -1,13 +1,11 @@
 <?php
 session_start();
-
 if (!isset($_SESSION['username'])) {
-    header("Location: ../login/login.html");
+    header('Location: /PORTALTRANSPARENCIAMUNI/login/login.html');
     exit();
 }
-?>
-<?php
-require_once '../connection/db.php';
+
+require_once __DIR__ . '/../connection/db.php';
 
 // Obtener el ID del documento a editar
 $id = $_GET['id'] ?? null;
@@ -47,7 +45,7 @@ $tipos = $stmt->fetchAll(PDO::FETCH_ASSOC);
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="css/editar_documento.css">
+    <link rel="stylesheet" href="../public/css/editar_documento.css">
 </head>
 
 <body>
