+++ admin/indexAdmin.php
@@ -1,10 +1,10 @@
 <?php
 session_start();
-
 if (!isset($_SESSION['username'])) {
-    header("Location: ../login/login.html");
+    header('Location: /PORTALTRANSPARENCIAMUNI/login/login.html');
     exit();
 }
+require_once __DIR__ . '/../connection/db.php';
 ?>
 <!DOCTYPE html>
 <html lang="es">
