+++ login/login.php
@@ -1,20 +1,36 @@
 <?php
 session_start();
+if (!isset($_SESSION['username'])) {
+    header('Location: /PORTALTRANSPARENCIAMUNI/login/login.html');
+    exit();
+}
+
 require_once '../connection/db.php';
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $username = $_POST['username'];
     $password = $_POST['password'];
 
-    $sql = "SELECT * FROM usuarios WHERE username = :username AND password = :password";
+    // Registrar:
+    $hashedPassword = password_hash($password, PASSWORD_DEFAULT);
+
+    $sql = "SELECT * FROM usuarios WHERE username = :username";
     $stmt = $pdo->prepare($sql);
     $stmt->bindParam(':username', $username);
-    $stmt->bindParam(':password', $password);
     $stmt->execute();
 
     if ($stmt->rowCount() == 1) {
-        $_SESSION['username'] = $username;
-        header('Location: ../public/indexAdmin.php');
+        $user = $stmt->fetch(PDO::FETCH_ASSOC);
+        $hashedPasswordFromDB = $user['password'];
+
+        // Validar:
+        if (password_verify($password, $hashedPasswordFromDB)) {
+            // Acceso concedido
+            $_SESSION['username'] = $username;
+            header('Location: /PORTALTRANSPARENCIAMUNI/admin/indexAdmin.php');
+        } else {
+            echo "Nombre de usuario o contraseña incorrectos.";
+        }
     } else {
         echo "Nombre de usuario o contraseña incorrectos.";
     }
